pyproject.toml
[project]
name = "auto-email-responder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-api-python-client",
  "google-auth-httplib2",
  "google-auth-oauthlib",
  "fastmcp",
  "beautifulsoup4",
  "openai>=1.97.0",
]

